refactor test to fit new game architecture
